mcp-server/requirements.txt
mcp[cli]
simple-salesforce
pandas
python-dotenv
pydantic
